requirements.txt
git+https://github.com/amazon-science/chronos-forecasting.git
datasets
pandas
numpy
chronos-forecasting
gluonts
torch
matplotlib
pyarrow
fastparquet